modifcations to make everything 64-bit compatible and still interoperate with 32-bit
git-svn-id: file:///mmt/REPOS/ao/trunk@43 91e47ee6-2bdd-4904-b37e-107d3e7cdac2

+++ data_server/DServ/DServ_Master.c
@@ -85,7 +85,7 @@ void *DServ_Master( void *Passed_Info)
   */
   for ( i=0; i<NUM_CONNECTIONS; i++) {
     Thread_Info[i].sockfd = -1;
-    Thread_Info[i].frameNumber = (long *)malloc(sizeof(long));
+    Thread_Info[i].frameNumber = (int *)malloc(sizeof(int));
   }
 
   /*=====================================================================
